Created DataAugmentation class, fixed KNN, NCC algorithms and more.

+++ algorithms/BaseClassifier.py
@@ -14,5 +14,5 @@ class BaseClassifier():
             Computes the "distance" metric between the given pokemons: {pokemonA}, {pokemonB}.
             The distance between 2 pokemons is the l2 norm of the matrix that results as the difference
             of the pixels of these images in every point
-        '''
-        return np.linalg.norm(pokemonA.getImagePixels() - pokemonB.getImagePixels())
+        '''        
+        return np.linalg.norm(pokemonA.getImagePixels().astype(np.int16) - pokemonB.getImagePixels().astype(np.int16))
